requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
PyPDF2
python-dotenv
aiofiles
